code-context/pyproject.toml
[project]
name = "code-context"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fast-plaid>=1.2.4.280",
    "fastmcp>=2.12.4",
    "gitpython>=3.1.45",
    "torch>=2.3.0",
    "Pillow>=10.3.0",
    "llama-cpp-python>=0.3.16",
    "huggingface-hub>=0.35.3",
    "onnxruntime-gpu>=1.23.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-language-pack>=0.10.0",
    "pathspec>=0.12.1",
    "polars>=1.34.0",
    "numpy~=1.0",
    "nbconvert>=7.16.6",
    "httpx>=0.28.1",
    "jq>=1.10.0",
    "faiss-gpu-cu12>=1.12.0",
]

[project.scripts]
serve = "code_context:main"
preindex-repos = "code_context.preindex_repos:preindex_github_repos"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
llama-cpp-python = { git = "https://github.com/abetlen/llama-cpp-python" }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.uv]
config-settings-package = { "llama-cpp-python" = { "cmake.args" = "-DGGML_CUDA=on" } }

[tool.pytest.ini_options]
asyncio_mode = "auto"
